Added high score functionality
Game waits for user to enter name, enter is bound to set name, once set game continues and it will use the enter binding for guessing.
Scores currently destroys main window and opens a seperate highscore window.
all scores are stored in a csv file, if the file does not exist, it first creates the file, then continues operation